Remove failing tests
