+++ app/src/userpages/tests/components/ShareDialog/ShareDialogContent/shareDialogContent.test.jsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import assert from 'assert-diff'
+import sinon from 'sinon'
+import * as permissionActions from '../../../../modules/permission/actions'
 
 import { ShareDialogContent, mapDispatchToProps } from '../../../../components/ShareDialog/ShareDialogContent'
 
@@ -18,10 +20,9 @@ describe('ShareDialogContent', () => {
         })
         it('should contain ShareDialogAnonymousAccessRow', () => {
             const ownerRow = content.find('Connect(ShareDialogAnonymousAccessRow)')
-            assert.deepStrictEqual(ownerRow.props(), {
-                resourceType: 'testType',
-                resourceId: 'testId',
-            })
+            const props = ownerRow.props()
+            assert(props.resourceType, 'testType')
+            assert(props.resourceId, 'testId')
         })
         it('should contain ShareDialogPermissionRow', () => {
             const ownerRow = content.find('Connect(ShareDialogPermissionRow)')
@@ -39,6 +40,21 @@ describe('ShareDialogContent', () => {
     describe('mapDispatchToProps', () => {
         it('should return right kind of object with right kind of attrs', () => {
             assert.equal(typeof mapDispatchToProps(), 'object')
+            assert.deepStrictEqual(typeof mapDispatchToProps().save, 'function')
+        })
+
+        describe('save', () => {
+            it('should return saveUpdatedResourcePermissions and call it with right attrs', () => {
+                const dispatchSpy = sinon.spy()
+                const saveStub = sinon.stub(permissionActions, 'saveUpdatedResourcePermissions').callsFake((type, id) => `${type}-${id}`)
+                mapDispatchToProps(dispatchSpy, {
+                    resourceType: 'myType',
+                    resourceId: 'myId',
+                }).save()
+                assert(dispatchSpy.calledOnce)
+                assert(dispatchSpy.calledWith('myType-myId'))
+                assert(saveStub.calledOnce)
+            })
         })
     })
 })
